+++ OAuth2.php
@@ -8,6 +8,7 @@
 namespace yii\authclient;
 
 use Yii;
+use yii\web\HttpException;
 
 /**
  * OAuth2 serves as a client for the OAuth 2 flow.
@@ -49,6 +50,15 @@ abstract class OAuth2 extends BaseOAuth
      * @var string token request URL endpoint.
      */
     public $tokenUrl;
+    /**
+     * @var boolean whether to use and validate auth 'state' parameter in authentication flow.
+     * If enabled - the opaque value will be generated and applied to auth URL to maintain
+     * state between the request and callback. The authorization server includes this value,
+     * when redirecting the user-agent back to the client.
+     * The option is used for preventing cross-site request forgery.
+     * @since 2.1
+     */
+    public $validateAuthState = true;
 
 
     /**
@@ -68,6 +78,12 @@ abstract class OAuth2 extends BaseOAuth
             $defaultParams['scope'] = $this->scope;
         }
 
+        if ($this->validateAuthState) {
+            $authState = $this->generateAuthState();
+            $this->setState('authState', $authState);
+            $defaultParams['state'] = $authState;
+        }
+
         return $this->composeUrl($this->authUrl, array_merge($defaultParams, $params));
     }
 
@@ -76,9 +92,19 @@ abstract class OAuth2 extends BaseOAuth
      * @param string $authCode authorization code, usually comes at $_GET['code'].
      * @param array $params additional request params.
      * @return OAuthToken access token.
+     * @throws HttpException on invalid auth state in case [[enableStateValidation]] is enabled.
      */
     public function fetchAccessToken($authCode, array $params = [])
     {
+        if ($this->validateAuthState) {
+            $authState = $this->getState('authState');
+            if (!isset($_REQUEST['state']) || empty($authState) || strcmp($_REQUEST['state'], $authState) !== 0) {
+                throw new HttpException(400, 'Invalid auth state parameter.');
+            } else {
+                $this->removeState('authState');
+            }
+        }
+
         $defaultParams = [
             'client_id' => $this->clientId,
             'client_secret' => $this->clientSecret,
@@ -145,9 +171,24 @@ abstract class OAuth2 extends BaseOAuth
     {
         $params = $_GET;
         unset($params['code']);
+        unset($params['state']);
         $params[0] = Yii::$app->controller->getRoute();
 
         return Yii::$app->getUrlManager()->createAbsoluteUrl($params);
+    }
+
+    /**
+     * Generates the auth state value.
+     * @return string auth state value.
+     * @since 2.1
+     */
+    protected function generateAuthState()
+    {
+        $baseString = get_class($this) . '-' . time();
+        if (Yii::$app->has('session')) {
+            $baseString .= '-' . Yii::$app->session->getId();
+        }
+        return hash('sha256', uniqid($baseString, true));
     }
 
     /**
@@ -195,10 +236,10 @@ abstract class OAuth2 extends BaseOAuth
     }
 
     /**
-     * Authenticates user directly by username/password pair, using 'password' grant type.
+     * Authenticates user directly by 'username/password' pair, using 'password' grant type.
      * @see https://tools.ietf.org/html/rfc6749#section-4.3
-     * @param string $username
-     * @param string $password
+     * @param string $username user name.
+     * @param string $password user password.
      * @param array $params additional request params.
      * @return OAuthToken access token.
      */